restructure API methods and rename files to a more meaniful name
