+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
 import addNewScore, { form } from './modules/addNewScore.js';
-import { getData } from './modules/APIMethods.js';
+import { getData } from './modules/API_GetData.js';
 import displayInHtml, { table } from './modules/displayInHtml.js';
 
 const refresh = document.querySelector('#refresh');
